move upload image function to firebaseRepository, fix upload image feature
